Refactor current env arch design components into DOW step 5
Move current env architectural design components into performance
requirements section (DOW step 5) used for the architectural design
card. The sections related to the DOW are now grouped together.
