+++ src/steps/05-PerformanceRequirements/DOW/ArchitecturalDesignDetails.vue
